% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinaryYScatterPlot.R
\name{BinaryYScatterPlot}
\alias{BinaryYScatterPlot}
\title{Plot a scatter plot of a binary variable.  xvar is the continuous independent variable and yvar is the dependent binary variable}
\usage{
BinaryYScatterPlot(frame, xvar, yvar, title = "")
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the independent column in frame}

\item{yvar}{name of the dependent (output or result to be modeled) column in frame}
}
\description{
Plot a scatter plot of a binary variable.  xvar is the continuous independent variable and yvar is the dependent binary variable
}

