% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{calcAUC}
\alias{calcAUC}
\title{calculate AUC.}
\usage{
calcAUC(modelPredictions, yValues)
}
\arguments{
\item{modelPredictions}{numeric predictions (not empty)}

\item{yValues}{logical truth (not empty, same lenght as model predictions)}
}
\value{
line graph, point graph, and area under curve
}
\description{
Based on:
 http://blog.revolutionanalytics.com/2016/08/roc-curves-in-two-lines-of-code.html
}
\details{
See also https://github.com/WinVector/sigr
}

