% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRPlot.R
\name{calcPR}
\alias{calcPR}
\title{calculate precision/recall curve.}
\usage{
calcPR(modelPredictions, yValues)
}
\arguments{
\item{modelPredictions}{numeric predictions (not empty)}

\item{yValues}{logical truth (not empty, same lenght as model predictions)}
}
\value{
line graph, point graph, and summaries
}
\description{
Based on:
 http://blog.revolutionanalytics.com/2016/08/roc-curves-in-two-lines-of-code.html
}
\details{
See also https://github.com/WinVector/sigr
}

