% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClevelandDotPlot.R
\name{ClevelandDotPlot}
\alias{ClevelandDotPlot}
\title{Plot a Cleveland dot plot.}
\usage{
ClevelandDotPlot(frm, xvar, title, ..., sort = -1, limit_n = NULL,
  stem = TRUE)
}
\arguments{
\item{frm}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}

\item{sort}{if TRUE sort data}

\item{limit_n}{if not NULL number of items to plot}

\item{stem}{if TRUE add stems/whiskers to plot}
}
\description{
assumes that xvar is a factor variable
sort < 0 sorts the factor levels in decreasing order (most frequent level first)
sort > 0 sorts the factor levels in increasing order (good when used in conjunction with coord_flip())
sort = 0 leaves the factor levels in "natural order" -- usually alphabetical
stem = FALSE will plot only the dots, without the stem to the y=0 line.
limit_n = NULL plots all the levels, N an integer limits to the top N most populous levels
}

