% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConditionalDistribution.R
\name{ConditionalDistributionPlot}
\alias{ConditionalDistributionPlot}
\title{Plot conditional density and conditional histogram.}
\usage{
ConditionalDistributionPlot(frame, xvar, truthVar, title, ..., breaks = 40)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}

\item{breaks}{number of breaks in histogram}
}
\description{
Plot conditional density and conditional histogram.
}

