% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebugPlot.R
\name{DebugPlot}
\alias{DebugPlot}
\title{Capture arguments of exception throwing plot for later debugging.}
\usage{
DebugPlot(plotFn, ...)
}
\arguments{
\item{plotFn}{}
}
\description{
Run and print plot.  Catch any errors and return either plot or arguments to re-run.
}
\examples{

d <- data.frame(x=1:5)
res = DebugPlot('PlotDistCountNormal',d,xvar='x','example')
res = DebugPlot('PlotDistCountNormal',d,xvar='xmisspelled','example')
if(!('ggplot' \%in\% class(res))) {
  print("error, re-run by running: do.call(res$plotFn,res$args)")
}

}

