% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiscreteDistribution.R
\name{DiscreteDistribution}
\alias{DiscreteDistribution}
\title{Plot distribution of a single continuous variable.}
\usage{
DiscreteDistribution(frm, xvar, title, ..., stem = TRUE)
}
\arguments{
\item{frm}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}

\item{stem}{if TRUE add whisker/stems to plot}
}
\description{
Plot distribution of a single continuous variable.
}

