% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoubleDensity.R
\name{DoubleDensityPlot}
\alias{DoubleDensityPlot}
\title{Plot two density plots conditioned on truthVar.}
\usage{
DoubleDensityPlot(frame, xvar, truthVar, title, ...)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}
}
\description{
Plot two density plots conditioned on truthVar.
}

