% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoubleHistogram.R
\name{DoubleHistogramPlot}
\alias{DoubleHistogramPlot}
\title{Plot two histograms conditioned on truthVar.}
\usage{
DoubleHistogramPlot(frame, xvar, truthVar, title, ..., breaks = 40)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}

\item{breaks}{breaks to pass to histogram}
}
\description{
Plot two histograms conditioned on truthVar.
}

