% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GainCurve.R
\name{GainCurvePlotC}
\alias{GainCurvePlotC}
\title{Plot the gain curve of a sort-order in horizontal orientation.}
\usage{
GainCurvePlotC(frame, xvar, costVar, truthVar, title, ...)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{costVar}{cost of each item (drives x-axis sum)}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}
}
\description{
Plot the gain curve of a sort-order in horizontal orientation.
}

