% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GainCurve.R
\name{GainCurvePlotWithNotation}
\alias{GainCurvePlotWithNotation}
\title{Take the standard WVPlots gain curve and add extra notation}
\usage{
GainCurvePlotWithNotation(frame, xvar, truthVar, title, gainx, labelfun, ...)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{gainx}{the point on the x axis corresponding to the desired label}

\item{labelfun}{a function to return a label for the marked point}

\item{...}{no unamed argument, added to force named binding of later arguments.}
}
\description{
Take the standard WVPlots gain curve and add extra notation
}
\examples{

set.seed(34903490)
x = rnorm(50)
y = 0.5*x^2 + 2*x + rnorm(length(x))
frm = data.frame(x=x,y=y,yC=y>=as.numeric(quantile(y,probs=0.8)))
frm$absY <- abs(frm$y)
frm$posY = frm$y > 0
frm$costX = 1
WVPlots::GainCurvePlot(frm, "x", "absY", title="Example Continuous Gain Curve")

}

