% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLogPlot.R
\name{LogLogPlot}
\alias{LogLogPlot}
\title{Plot a trend on log-log paper.}
\usage{
LogLogPlot(frame, xvar, yvar, title, ...)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the independent (input or model) column in frame}

\item{yvar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unnamed argument, added to force named binding of later arguments.}
}
\description{
This plot is intended for ploting functions that are observed costs
or durations as a function of problem size.  In this case we expect
the ideal or expected cost function to be non-decreasing.
Any negative trends are assumed to arise from the noise model.
The graph is specialized to compare non-decreasing linear and
non-decreasing quadratic growth.
}
\examples{

set.seed(5326)
frm = data.frame(x = 1:20)
frm$y <- 5 + frm$x + 0.2 * frm$x * frm$x + 0.1*abs(rnorm(nrow(frm)))
WVPlots::LogLogPlot(frm, "x", "y", title="Example Trend")

}
