% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdPlot.R
\name{MetricPairPlot}
\alias{MetricPairPlot}
\title{Plot the relationship between two metrics.}
\usage{
MetricPairPlot(
  frame,
  xvar,
  truthVar,
  title,
  ...,
  x_metric = "false_positive_rate",
  y_metric = "true_positive_rate",
  truth_target = TRUE
)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the independent (input or model) column in frame}

\item{truthVar}{name of the column to be predicted}

\item{title}{title to place on plot}

\item{...}{no unarmed argument, added to force named binding of later arguments.}

\item{x_metric}{metric to be plotted, allowed: 'threshold', 'count', 'fraction', 'precision', 'true_positive_rate', 'false_positive_rate', 'true_negative_rate', 'false_negative_rate', 'recall', 'sensitivity', 'specificity', 'cdf', 'pdf'.}

\item{y_metric}{metric to be plotted, allowed: 'threshold', 'count', 'fraction', 'precision', 'true_positive_rate', 'false_positive_rate', 'true_negative_rate', 'false_negative_rate', 'recall', 'sensitivity', 'specificity', 'cdf', 'pdf'.}

\item{truth_target}{truth value considered to be positive.}
}
\description{
Plot the relationship between two metrics.
}
\examples{

# data with two different regimes of behavior
d <- rbind(
  data.frame(
    x =  rnorm(1000),
    y = sample(c(TRUE, FALSE), prob = c(0.02, 0.98), size = 1000, replace = TRUE)),
  data.frame(
    x =  rnorm(200) + 5,
    y = sample(c(TRUE, FALSE), size = 200, replace = TRUE))
)
# sensitivity/1-specificity examples
ThresholdPlot(d, 'x', 'y', 'positivity rates, and data distribution',
   metrics = c('sensitivity', 'specificity', 'pdf'))
MetricPairPlot(d, 'x', 'y', 'ROC equivalent')
ROCPlot(d, 'x', 'y', TRUE, 'ROC example')
# precision/recall examples
ThresholdPlot(d, 'x', 'y', 'precision/recall, and data distribution',
   metrics = c('recall', 'precision', 'pdf'))
MetricPairPlot(d, 'x', 'y', 'recall/precision', x_metric = 'recall', y_metric = 'precision')

}
