% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairPlot.R
\name{PairPlot}
\alias{PairPlot}
\title{Build a pair plot}
\usage{
PairPlot(d, meas_vars, title, ..., group_var = NULL, palette = "Dark2")
}
\arguments{
\item{d}{data frame}

\item{meas_vars}{the variables to be plotted}

\item{title}{plot title}

\item{...}{not used, forces later arguments to bind by name}

\item{group_var}{variable for grouping and colorcoding}

\item{palette}{name of a brewer palette (NULL for ggplot2 default coloring)}
}
\value{
a ggplot2 pair plot
}
\description{
Build a pair plot
}
\examples{

PairPlot(iris, colnames(iris)[1:4], "Example plot", group_var = "Species")

# no color-coding
PairPlot(iris, colnames(iris)[1:4], "Example plot")

}
