% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHistoryPlot.R
\name{plot_Keras_fit_trajectory}
\alias{plot_Keras_fit_trajectory}
\title{Plot the trajectory of a Keras model fit.}
\usage{
plot_Keras_fit_trajectory(d, title, ..., epoch_name = "epoch",
  pick_metric = "minus binary cross entropy", discount_rate = 0.1)
}
\arguments{
\item{d}{data frame to get values from.}

\item{title}{character title for plot.}

\item{...}{force later arguments to be bound by name}

\item{epoch_name}{name for epoch or trajectory column.}

\item{pick_metric}{character metric to maximize.}

\item{discount_rate}{numeric what fraction of over-fit to subtract from validation performance.}
}
\value{
ggplot2 plot
}
\description{
Plot a history of fit over a trajetory of times.  Example below gives
a fit plot for a history report from Keras R package.
Please see \url{http://winvector.github.io/FluidData/PlotExample/KerasPerfPlot.html}
for some details.
}
\examples{

# example data (from Keras)
d <- data.frame(
  val_loss = c(0.3769818, 0.2996994, 0.2963943, 0.2779052, 0.2842501),
  val_acc  = c(0.8722000, 0.8895000, 0.8822000, 0.8899000, 0.8861000),
  loss     = c(0.5067290, 0.3002033, 0.2165675, 0.1738829, 0.1410933),
  acc      = c(0.7852000, 0.9040000, 0.9303333, 0.9428000, 0.9545333) )

plt <- plot_Keras_fit_trajectory(
  d,
  title = "model performance by epoch, dataset, and measure")

suppressWarnings(print(plt)) # too few points for loess

}
\seealso{
\link{plot_fit_trajectory}
}
