% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{plotlyROC}
\alias{plotlyROC}
\title{Use \code{plotly} to produce a ROC plot.}
\usage{
plotlyROC(d, predCol, outcomeCol, outcomeTarget, title, ...)
}
\arguments{
\item{d}{dataframe}

\item{predCol}{name of column with numeric predictions}

\item{outcomeCol}{name of column with truth}

\item{outcomeTarget}{value considred true}

\item{title}{character title for plot}

\item{...}{no unnamed argument, added to force named binding of later arguments.}
}
\value{
plotly plot
}
\description{
Use \code{plotly} to produce a ROC plot.
}
\examples{


d <- data.frame(x= 1:5, y= c(TRUE, FALSE , TRUE, TRUE, TRUE))
plotlyROC(d, 'x', 'y', TRUE, 'example plot')

}
