% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ScatterHist.R
\name{ScatterHist}
\alias{ScatterHist}
\title{Plot a scatter plot.  xvar is the independent variable (input or model) and yvar is the dependent variable}
\usage{
ScatterHist(frame, xvar, yvar, smoothmethod = "auto", title = "",
  annot_size = 5, minimal_labels = TRUE, binwidth_x = NULL,
  binwidth_y = NULL, adjust_x = 1, adjust_y = 1)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{yvar}{name of the dependent (output or result to be modeled) column in frame}

\item{smoothmethod}{(optional) one of 'auto' (the default), 'lm', or 'identity'.  If smoothmethod is 'auto' or 'lm' a smoothing curve or line (respectively) is added and R-squared of the best linear fit of xvar to yvar is reported.  If smoothmethod is 'identity' then the y=x line is added and the R-squared of xvar to yvar (without the linear transform used in the other smoothmethod modes) is reported.}
}
\description{
Plot a scatter plot.  xvar is the independent variable (input or model) and yvar is the dependent variable
}

