% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ScatterPlot.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{Plot a scatter plot.  xvar is the independent variable (input or model) and yvar is the dependent variable}
\usage{
ScatterPlot(frame, xvar, yvar, smoothmethod = "auto", title = "",
  annot_size = 5, minimal_labels = TRUE, binwidth_x = NULL,
  binwidth_y = NULL, adjust_x = 1, adjust_y = 1)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}
}
\description{
Plot a scatter plot.  xvar is the independent variable (input or model) and yvar is the dependent variable
}

