% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShadedDensity.R
\name{ShadedDensity}
\alias{ShadedDensity}
\title{Plot the distribution of a variable with a tail shaded}
\usage{
ShadedDensity(frame, xvar, threshold, title, ..., tail = "left")
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the variable to be density plotted}

\item{threshold}{boundary value for the tail}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}

\item{tail}{which tail to shade, 'left' (default) or 'right'}
}
\description{
Plot the distribution of a variable with a tail shaded
}
\examples{
set.seed(34903490)
x = rnorm(50)
frm = data.frame(x=x)
threshold = -1
WVPlots::ShadedDensity(frm, "x", title="Example shaded density plot")
}

