% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdPlot.R
\name{ThresholdPlot}
\alias{ThresholdPlot}
\title{Plot distribution of metrics as a function of being greater than or equal to thresholds.}
\usage{
ThresholdPlot(
  frame,
  xvar,
  truthVar,
  title,
  ...,
  metrics = c("sensitivity", "specificity"),
  truth_target = TRUE
)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the independent (input or model) column in frame}

\item{truthVar}{name of the column to be predicted}

\item{title}{title to place on plot}

\item{...}{no unarmed argument, added to force named binding of later arguments.}

\item{metrics}{metrics to be computed, allowed: 'threshold', 'count', 'fraction', 'precision', 'true_positive_rate', 'false_positive_rate', 'true_negative_rate', 'false_negative_rate', 'recall', 'sensitivity', 'specificity', 'cdf', 'pdf', 'positive_rate'.}

\item{truth_target}{truth value considered to be positive.}
}
\description{
Plot distribution of metrics as a function of being greater than or equal to thresholds.
}
\examples{

d <- data.frame(
 x =  rnorm(1000),
 y = sample(c(TRUE, FALSE), size = 1000, replace = TRUE))
ThresholdPlot(d, 'x', 'y', 'positivity rates, and data distribution',
   metrics = c('sensitivity', 'specificity', 'pdf'))
ThresholdPlot(d, 'x', 'y', 'precision/recall, and data distribution',
   metrics = c('recall', 'precision', 'pdf'))
MetricPairPlot(d, 'x', 'y', 'ROC equivalent')
ROCPlot(d, 'x', 'y', TRUE, 'ROC example')

}
