% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVPlots.R
\docType{package}
\name{WVPlots}
\alias{WVPlots}
\alias{WVPlots-package}
\title{WVPlots: Common Plots for Analysis}
\description{
Example 'ggplot2' plots we have found useful, under a standardized
calling interface.
}
\details{
To learn more about replyr, please start with the vignette:
\code{vignette('replyr','replyr')}

For more information:
\itemize{
  \item \code{vignette(package='WVPlots')}
  \item \code{RShowDoc('WVPlots_examples',package='WVPlots')}
  \item Website: \url{https://github.com/WinVector/WVPlots} }
}

