% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinaryYScatterPlot.R
\name{BinaryYScatterPlot}
\alias{BinaryYScatterPlot}
\title{Plot a scatter plot of a binary variable.  xvar is the continuous independent variable and yvar is the dependent binary variable}
\usage{
BinaryYScatterPlot(frame, xvar, yvar, title, ..., se = FALSE,
  use_glm = FALSE)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the independent column in frame}

\item{yvar}{name of the dependent (output or result to be modeled) column in frame}

\item{se}{if TRUE, add error bars (defaults to FALSE). Ignored if useGLM is TRUE}

\item{use_glm}{if TRUE, "smooths" with a one-variable logistic regression (defaults to FALSE)}
}
\description{
Plot a scatter plot of a binary variable.  xvar is the continuous independent variable and yvar is the dependent binary variable
}

