% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalDistribution.R
\name{ConditionalDistributionPlot}
\alias{ConditionalDistributionPlot}
\title{Plot conditional density and conditional histogram.}
\usage{
ConditionalDistributionPlot(frame, xvar, truthVar, title, ..., breaks = 40)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the indepement (input or model) column in frame}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unamed argument, added to force named binding of later arguments.}

\item{breaks}{number of breaks in histogram}
}
\description{
Plot conditional density and conditional histogram.
}
\examples{

set.seed(34903490)
x = rnorm(50)
y = 0.5*x^2 + 2*x + rnorm(length(x))
frm = data.frame(x=x,y=y,yC=y>=as.numeric(quantile(y,probs=0.8)))
frm$absY <- abs(frm$y)
frm$posY = frm$y > 0
frm$costX = 1
WVPlots::ConditionalDistributionPlot(frm, "x", "yC", title="Example conditional distribution plot")

}

