% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleDensity.R
\name{DoubleDensityPlot}
\alias{DoubleDensityPlot}
\title{Plot two density plots conditioned on an outcome variable.}
\usage{
DoubleDensityPlot(frame, xvar, truthVar, title, ..., truth_target = NULL)
}
\arguments{
\item{frame}{data frame to get values from}

\item{xvar}{name of the independent (input or model) column in frame}

\item{truthVar}{name of the dependent (output or result to be modeled) column in frame}

\item{title}{title to place on plot}

\item{...}{no unnamed argument, added to force named binding of later arguments.}

\item{truth_target}{if not NULL compare to this scalar value.}
}
\description{
Plot two density plots conditioned on a binary outcome variable.
}
\details{
The use case for this visualization is to plot the distribution of a predictive model score
(usually the predicted probability
of a desired outcome) conditioned on the actual outcome. However, you can use it to compare the distribution of any
numerical quantity conditioned on a binary feature. See the examples.

The plot will degrade gracefully in degenerate conditions, for example when only
one category is present.
}
\examples{

mpg = ggplot2::mpg
mpg$trans = gsub("\\\\(.*$", '', mpg$trans)
WVPlots::DoubleDensityPlot(mpg, "cty", "trans", "City driving mpg by transmission type")

set.seed(34903490)
x = rnorm(50)
y = 0.5*x^2 + 2*x + rnorm(length(x))
frm = data.frame(score=x,
   truth=(y>=as.numeric(quantile(y,probs=0.8))),
   stuck=TRUE,
   rare=FALSE)
frm[1,'rare'] = TRUE
WVPlots::DoubleDensityPlot(frm, "score", "truth", title="Example double density plot")
WVPlots::DoubleDensityPlot(frm, "score", "stuck", title="Example double density plot")
WVPlots::DoubleDensityPlot(frm, "score", "rare", title="Example double density plot")

}
