% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdPlot.R
\name{ThresholdPlot}
\alias{ThresholdPlot}
\title{Plot distribution of metrics as a function of being greater than or equal to thresholds.}
\usage{
ThresholdPlot(
  frame,
  xvar,
  truthVar,
  title,
  ...,
  metrics = c("sensitivity", "specificity"),
  truth_target = TRUE
)
}
\arguments{
\item{xvar}{name of the independent (input or model) column in frame}

\item{truthVar}{name of the column to be predicted}

\item{title}{title to place on plot}

\item{...}{no unarmed argument, added to force named binding of later arguments.}

\item{metrics}{metrics to be computed, allowed: 'threshold', 'count', 'fraction', 'precision', 'true_positive_rate', 'false_positive_rate', 'true_negative_rate', 'false_negative_rate', 'recall', 'sensitivity', 'specificity'.}

\item{truth_target}{truth value considered to be positive.}

\item{frm}{data frame to get values from}
}
\description{
Plot distribution of metrics as a function of being greater than or equal to thresholds.
}
\examples{

d <- data.frame(
 x =  c(1, 2, 3, 4, 5),
  y = c(FALSE, FALSE, TRUE, TRUE, FALSE))
ThresholdPlot(d, 'x', 'y', 'example plot')

}
