\name{WVPlots-package}
\alias{WVPlots-package}
\alias{WVPlots}
\docType{package}
\title{
WVPlots: example plots
}
\description{
WVPlots is a package of common useful plots on top of ggplot2.
}
\details{
\tabular{ll}{
Package: \tab WVPlots\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2017-01-15\cr
License: \tab GNU General Public License version 3\cr
}
}
\author{
Nina Zumel and John Mount

Maintainer: <contact@win-vector.com>
}


\keyword{ package }
\seealso{
For more information:
\itemize{
  \item \code{vignette(package='WVPlots')}
  \item \code{RShowDoc('WVPlots_examples',package='WVPlots')}
  \item Website: \url{https://github.com/WinVector/WVPlots}
}
}

