% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVPlots.R
\docType{package}
\name{WVPlots}
\alias{WVPlots}
\alias{WVPlots-package}
\title{WVPlots: Common Plots for Analysis}
\description{
Select data analysis plots, under a standardized calling interface implemented
on top of \code{ggplot2} and \code{plotly}.
Plots of interest include: \code{ROC}, gain curve, scatter plot with marginal distributions,
conditioned scatter plot with marginal densities.
box and stem with matching theoretical distribution, density with matching theoretical distribution.
}
\details{
For more information:
\itemize{
  \item \code{vignette(package='WVPlots')}
  \item \code{RShowDoc('WVPlots_examples',package='WVPlots')}
  \item Website: \url{https://github.com/WinVector/WVPlots} }
}
